move index watching code to lib/index-static-watcher.js.
